requirements.txt
slack-bolt
apscheduler
python-dotenv
google-generativeai
pinecone
langchain-text-splitters
pypdf
groq
faiss-cpu
sentence-transformers
numpy
python-docx
pandas
openpyxl
